% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROHMOplotter.R
\name{ROHMOplotter}
\alias{ROHMOplotter}
\title{Plot genome-wide BAF, LRR and Hr computed by ROHMOseeker}
\usage{
ROHMOplotter(hetseries, dir)
}
\arguments{
\item{hetseries}{List obtained with ROHMOseeker containing detected ROH, mCA and plot parameters}

\item{dir}{Direcory where the plot file shoud be created}
}
\description{
Plot genome-wide BAF, LRR and Hr computed by ROHMOseeker
}
