% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROHMOseeker.R
\name{ROHMOseeker}
\alias{ROHMOseeker}
\title{Detect ROH and mCA regions from BAF and LRR}
\usage{
ROHMOseeker(
  sub,
  bafsubs,
  lrrsubs,
  winhet = 1000,
  winhetrel = 500,
  winlrr = 1000,
  plot = FALSE
)
}
\arguments{
\item{sub}{name of the sample to process}

\item{bafsubs}{Data frame with BAF of the sample to process for each probe. First to third columns should contain probe names, chromosome and position.
Column name of the column containing the BAF values should be the same as specified in 'sub'. Extra columns are ignored.}

\item{lrrsubs}{Data frame with LRR of the sample to process for each probe. First to third columns should contain probe names, chromosome and position.
Column name of the column containing the LRR values should be the same as specified in 'sub'. Extra columns are ignored.}

\item{winhet}{Window size for BAF}

\item{winhetrel}{Window size for Hr}

\item{winlrr}{Window size for LRR}

\item{plot}{If \code{TRUE}, then plot parameters are returned in the output to use with \code{\link{ROHMOplotter}()}}
}
\value{
A list with ROH and mCA detected coordinates. If specified, also plot parameters
}
\description{
Detect ROH and mCA regions from BAF and LRR
}
